refactor: icon 배럴로 내보내기
