+++ src/assets/icons/index.tsx
@@ -1,0 +1,3 @@
+export { ReactComponent as ClearIcon } from '@assets/icons/clear.svg';
+export { ReactComponent as CalendarIcon } from '@assets/icons/calendar.svg';
+export { ReactComponent as IOSRightArrowIcon } from '@assets/icons/iOSRightArrow.svg';
